Refactored the TCP / UDP implementation (almost ready for release, but I still want to check if I can go further).
