+++ Common/Net/IP/IPHelper.NativeMethods.cs
@@ -3,11 +3,9 @@ using System.Runtime.InteropServices;
 
 namespace Wokhan.WindowsFirewallNotifier.Common.Net.IP;
 
-public partial class IPHelper
+internal static partial class NativeMethods
 {
-    protected static partial class NativeMethods
-    {
-        [LibraryImport("kernel32.dll", SetLastError = false)]
-        internal static partial void RtlZeroMemory(IntPtr dest, uint size);
-    }
+    [LibraryImport("kernel32.dll", SetLastError = false)]
+    internal static partial void RtlZeroMemory(IntPtr dest, uint size);
 }
+
