--- Common/Net/IP/IPHelper.TCPIP_OWNER_MODULE_BASIC_INFO.cs
@@ -1,14 +1,0 @@
-﻿using System;
-using System.Runtime.InteropServices;
-
-namespace Wokhan.WindowsFirewallNotifier.Common.Net.IP;
-
-public abstract partial class IPHelper
-{
-    [StructLayout(LayoutKind.Sequential)]
-    internal struct TCPIP_OWNER_MODULE_BASIC_INFO
-    {
-        public IntPtr p1;
-        public IntPtr p2;
-    }
-}
