--- Common/Net/IP/IPHelper.TCPIP_OWNER_MODULE_INFO_CLASS.cs
@@ -1,9 +1,0 @@
-﻿namespace Wokhan.WindowsFirewallNotifier.Common.Net.IP;
-
-public abstract partial class IPHelper
-{
-    internal enum TCPIP_OWNER_MODULE_INFO_CLASS
-    {
-        TCPIP_OWNER_MODULE_INFO_BASIC
-    }
-}
