+++ Common/Net/IP/TCP/TCP4/MIB_TCPTABLE_OWNER_MODULE.cs
@@ -1,13 +1,11 @@
 ﻿using System.Runtime.InteropServices;
 
-namespace Wokhan.WindowsFirewallNotifier.Common.Net.IP.TCP;
 
-public partial class TCPHelper
+namespace Wokhan.WindowsFirewallNotifier.Common.Net.IP.TCP.TCP4;
+
+[StructLayout(LayoutKind.Sequential)]
+internal struct MIB_TCPTABLE_OWNER_MODULE
 {
-    [StructLayout(LayoutKind.Sequential)]
-    internal struct MIB_TCPTABLE_OWNER_MODULE
-    {
-        public uint NumEntries;
-        public MIB_TCPROW_OWNER_MODULE FirstEntry;
-    }
+    public uint NumEntries;
+    public MIB_TCPROW_OWNER_MODULE FirstEntry;
 }
