--- Common/Net/IP/TCP/TCPHelper.TCP_BOOLEAN_OPTIONAL.cs
@@ -1,12 +1,0 @@
-﻿namespace Wokhan.WindowsFirewallNotifier.Common.Net.IP.TCP;
-
-public partial class TCPHelper
-{
-    internal enum TCP_BOOLEAN_OPTIONAL
-    {
-        TcpBoolOptDisabled = 0,
-        TcpBoolOptEnabled = 1,
-        TcpBoolOptUnchanged = -1
-    }
-
-}
