--- Common/Net/IP/TCP/TCPHelper.TCP_TABLE_CLASS.cs
@@ -1,18 +1,0 @@
-﻿namespace Wokhan.WindowsFirewallNotifier.Common.Net.IP.TCP;
-
-public partial class TCPHelper
-{
-    protected enum TCP_TABLE_CLASS
-    {
-        TCP_TABLE_BASIC_LISTENER,
-        TCP_TABLE_BASIC_CONNECTIONS,
-        TCP_TABLE_BASIC_ALL,
-        TCP_TABLE_OWNER_PID_LISTENER,
-        TCP_TABLE_OWNER_PID_CONNECTIONS,
-        TCP_TABLE_OWNER_PID_ALL,
-        TCP_TABLE_OWNER_MODULE_LISTENER,
-        TCP_TABLE_OWNER_MODULE_CONNECTIONS,
-        TCP_TABLE_OWNER_MODULE_ALL
-    }
-
-}
