+++ Common/Net/IP/TCP/TCP_BOOLEAN_OPTIONAL.cs
@@ -1,0 +1,8 @@
+﻿namespace Wokhan.WindowsFirewallNotifier.Common.Net.IP.TCP;
+
+internal enum TCP_BOOLEAN_OPTIONAL
+{
+    TcpBoolOptDisabled = 0,
+    TcpBoolOptEnabled = 1,
+    TcpBoolOptUnchanged = -1
+}
