+++ Common/Net/IP/TCP/TCP_ESTATS_BANDWIDTH_ROD_v0.cs
@@ -1,14 +1,11 @@
 ﻿namespace Wokhan.WindowsFirewallNotifier.Common.Net.IP.TCP;
 
-public partial class TCPHelper
+public struct TCP_ESTATS_BANDWIDTH_ROD_v0
 {
-    public struct TCP_ESTATS_BANDWIDTH_ROD_v0
-    {
-        public ulong OutboundBandwidth;
-        public ulong InboundBandwidth;
-        public ulong OutboundInstability;
-        public ulong InboundInstability;
-        public bool OutboundBandwidthPeaked;
-        public bool InboundBandwidthPeaked;
-    }
+    public ulong OutboundBandwidth;
+    public ulong InboundBandwidth;
+    public ulong OutboundInstability;
+    public ulong InboundInstability;
+    public bool OutboundBandwidthPeaked;
+    public bool InboundBandwidthPeaked;
 }
