+++ Common/Net/IP/TCPIP_OWNER_MODULE_INFO_CLASS.cs
@@ -1,0 +1,6 @@
+﻿namespace Wokhan.WindowsFirewallNotifier.Common.Net.IP;
+
+internal enum TCPIP_OWNER_MODULE_INFO_CLASS
+{
+    TCPIP_OWNER_MODULE_INFO_BASIC
+}
