+++ Common/Net/IP/UDP/UDP4/MIB_UDPROW_OWNER_MODULE.cs
@@ -4,36 +4,36 @@ using System.Runtime.InteropServices;
 
 using Wokhan.WindowsFirewallNotifier.Common.Net.IP.TCP;
 
-namespace Wokhan.WindowsFirewallNotifier.Common.Net.IP.UDP6;
+namespace Wokhan.WindowsFirewallNotifier.Common.Net.IP.UDP.UDP4;
 
 [StructLayout(LayoutKind.Sequential)]
-internal struct MIB_UDP6ROW_OWNER_MODULE : IConnectionOwnerInfo
+internal struct MIB_UDPROW_OWNER_MODULE : IConnectionOwnerInfo
 {
-    [MarshalAs(UnmanagedType.ByValArray, SizeConst = 16)]
-    internal byte[] _localAddress;
-    internal uint LocalScopeId;
-    [MarshalAs(UnmanagedType.ByValArray, SizeConst = 4)]
+    [MarshalAs(UnmanagedType.ByValArray, SizeConst = 4)]//, FieldOffset(0)]
+    internal byte[] _localAddr;
+    [MarshalAs(UnmanagedType.ByValArray, SizeConst = 4)]//, FieldOffset(4)]
     internal byte[] _localPort;
+    //[FieldOffset(8)]
     internal uint _owningPid;
     //[FieldOffset(16)]
     internal long _creationTime;
     //[FieldOffset(24)]
     //public int SpecificPortBind;
     //[FieldOffset(24)]
-    internal int Flags;
+    internal int _flags;
     [MarshalAs(UnmanagedType.ByValArray, SizeConst = 16)] //FieldOffset(32), 
-    internal ulong[] OwningModuleInfo;
+    internal ulong[] _owningModuleInfo;
 
     public byte[] RemoteAddrBytes => Array.Empty<byte>();
     public ConnectionStatus State => ConnectionStatus.NOT_APPLICABLE;
     public uint OwningPid => _owningPid;
-    public string LocalAddress => IPHelper.GetRealAddress(_localAddress);
+    public string LocalAddress => IPHelper.GetAddressAsString(_localAddr);
     public int LocalPort => IPHelper.GetRealPort(_localPort);
-    public Owner? OwnerModule => UDP6Helper.GetOwningModuleUDP6(this);
+    public Owner? OwnerModule => IPHelper.GetOwningModuleInternal(NativeMethods.GetOwnerModuleFromUdpEntry, this);
     public string Protocol => "UDP";
+    public DateTime? CreationTime => _creationTime == 0 ? null : DateTime.FromFileTime(_creationTime);
     public string RemoteAddress => string.Empty;
     public int RemotePort => -1;
-    public DateTime? CreationTime => _creationTime == 0 ? (DateTime?)null : DateTime.FromFileTime(_creationTime);
     public bool IsLoopback => IPAddress.IsLoopback(IPAddress.Parse(RemoteAddress));
 
     public ITcpRow ToTcpRow()
