--- Common/Net/IP/UDP/UDPHelper.UDP_TABLE_CLASS.cs
@@ -1,11 +1,0 @@
-﻿namespace Wokhan.WindowsFirewallNotifier.Common.Net.IP.UDP;
-
-public partial class UDPHelper
-{
-    protected enum UDP_TABLE_CLASS
-    {
-        UDP_TABLE_BASIC,
-        UDP_TABLE_OWNER_PID,
-        UDP_TABLE_OWNER_MODULE
-    }
-}
