+++ Common/Net/IP/UDP/UDP_TABLE_CLASS.cs
@@ -1,0 +1,8 @@
+﻿namespace Wokhan.WindowsFirewallNotifier.Common.Net.IP.UDP;
+
+internal enum UDP_TABLE_CLASS
+{
+    UDP_TABLE_BASIC,
+    UDP_TABLE_OWNER_PID,
+    UDP_TABLE_OWNER_MODULE
+}
