--- Common/Net/IP/UDP6/UDP6Helper.MIB_UDP6TABLE_OWNER_MODULE.cs
@@ -1,13 +1,0 @@
-﻿using System.Runtime.InteropServices;
-
-namespace Wokhan.WindowsFirewallNotifier.Common.Net.IP.UDP6;
-
-public partial class UDP6Helper
-{
-    [StructLayout(LayoutKind.Sequential)]
-    private struct MIB_UDP6TABLE_OWNER_MODULE
-    {
-        public uint NumEntries;
-        public MIB_UDP6ROW_OWNER_MODULE FirstEntry;
-    }
-}
